Removed some warnings. Moved the configuration to a separate folder making it easier to build for different environments.
git-svn-id: http://ikube.googlecode.com/svn/trunk@800 89a0fb20-889e-aefa-90f6-7c296b20d5cf

+++ code/core/src/main/java/ikube/IndexEngine.java
@@ -57,7 +57,7 @@ public class IndexEngine implements IIndexEngine, IListener {
 			for (IAction<IndexContext<?>, Boolean> action : actions) {
 				try {
 					action.execute(indexContext);
-					ThreadUtilities.sleep((long) Math.abs(random.nextLong()) % 3000);
+					ThreadUtilities.sleep(Math.abs(random.nextLong()) % 3000l);
 				} catch (Exception e) {
 					LOGGER.error("Exception executing action : " + action, e);
 				}
